backend/requirements.txt
PySide6>=6.0.0
ollama>=0.4.7
scikit-learn>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
Flask>=2.0.0
Flask-SQLAlchemy>=3.0.0
PyJWT>=2.0.0
cryptography>=3.0.0